Refact folder structure
이전 Refactoring 파일의 연장선
RegisterProgress에 불필요하게 있던 파일 삭제
RegisterProgress, SelectBox, RegisterTermSelector 이동
