+++ client/src/pages/Register/index.jsx
@@ -2,7 +2,7 @@ import React, { useState } from "react"
 import styled from "styled-components"
 
 import Footer from "../../components/Atoms/Footer"
-import Progress from "../../components/RegisterProgress"
+import Progress from "../../components/Organisim/RegisterProgress"
 import SelectCategory from "./template/SelectCategory"
 import InsertInfo from "./template/InsertInfo"
 import Complete from "./template/Complete"
